requirements.txt
asgiref==3.8.1
atomicwrites==1.4.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==2.0.12
colorama==0.4.6
Django==5.1.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.10
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==0.13.1
py==1.11.0
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2025.1
requests==2.26.0
sqlparse==0.5.3
toml==0.10.2
tzdata==2025.1
urllib3==1.26.20
